requirements.txt
amqp==1.4.6
anyjson==0.3.3
argh==0.26.1
billiard==3.3.0.20
celery==3.1.18
codegen==1.0
Django==1.8.2
django-celery==3.1.16
django-filebrowser==3.5.7
django-grappelli==2.5.7
django-gunicorn==0.1.1
django-supervisor==0.3.2
gunicorn==19.3.0
image==1.3.9
kombu==3.0.26
meld3==1.0.2
pathtools==0.1.2
patterns==0.3
pdfminer==20110515
Pillow==2.8.1
pytesseract==0.1.6
python-magic==0.4.6
pytz==2015.4
PyYAML==3.11
slate==0.3
supervisor==3.1.3
watchdog==0.8.3
wheel==0.24.0
